Proper functionality with Snow and polymorphism
I fiddled with the Snow class and got it functioning to a point that I
am satisfied with as far as the speed of the vertical descent and the
horizontal drift.
I also added implementation for the use of the
polymorphic capability of the Particles class through the main file
(assn1.cpp). This also functions well by taking inputs from the
keyboard and changes the particles displayed immediately.

To work on:
- fire/flame class' movement code
- adding some sort of instructions
- ripple effect? (ultimate goals)